push et ajout de 0-O et 2-O

+++ 0-bubble_sort.c
@@ -11,9 +11,9 @@ void bubble_sort(int *array, size_t size)
 
 {
 
-long unsigned int tmp = 0;
-long unsigned int i;
-long unsigned int j = 0;
+int tmp = 0;
+int i;
+int j = 0;
 
 
 while (array[j] != array[size - 1])
@@ -24,8 +24,8 @@ i = 0;
 		if (array[i] > array[i + 1])
 		{
 			tmp = array[i];
-			array[i] = array[i+1];
-			array[i+1] = tmp;
+			array[i] = array[i + 1];
+			array[i + 1] = tmp;
 			print_array(array, size);
 		}
 		i++;
